fix scrolls

+++ src/main/java/com/sun/manager/forms/admin/MainAdminController.java
@@ -493,33 +493,39 @@ public class MainAdminController extends ScrollPane implements Initializable {
                 @Override
                 public void handle(MouseEvent mouseEvent) {
                     BaseSolariumData data = tableVert.getSelectionModel().getSelectedItem();
-                    if (data.isSaved()) {
-                        solariumService.deleteSolariumData(data, 1L);
+                    if (data != null) {
+                        if (data.isSaved()) {
+                            solariumService.deleteSolariumData(data, 1L);
+                        }
+                        vertData.remove(data);
+                        vertData.add((BaseSolariumData) BlankItem.generateBlankItem(1L));
                     }
-                    vertData.remove(data);
-                    vertData.add((BaseSolariumData)BlankItem.generateBlankItem(1L));
                 }
             });
             del2.setOnMouseClicked(new EventHandler<MouseEvent>() {
                 @Override
                 public void handle(MouseEvent mouseEvent) {
                     BaseSolariumData data = tableGreen.getSelectionModel().getSelectedItem();
-                    if (data.isSaved()) {
-                        solariumService.deleteSolariumData(data, 2L);
+                    if (data != null) {
+                        if (data.isSaved()) {
+                            solariumService.deleteSolariumData(data, 2L);
+                        }
+                        greenData.remove(data);
+                        greenData.add((BaseSolariumData) BlankItem.generateBlankItem(1L));
                     }
-                    greenData.remove(data);
-                    greenData.add((BaseSolariumData)BlankItem.generateBlankItem(1L));
                 }
             });
             del3.setOnMouseClicked(new EventHandler<MouseEvent>() {
                 @Override
                 public void handle(MouseEvent mouseEvent) {
                     BaseSolariumData data = tableBlue.getSelectionModel().getSelectedItem();
-                    if (data.isSaved()) {
-                        solariumService.deleteSolariumData(data, 3L);
+                    if (data != null) {
+                        if (data.isSaved()) {
+                            solariumService.deleteSolariumData(data, 3L);
+                        }
+                        blueData.remove(data);
+                        blueData.add((BaseSolariumData) BlankItem.generateBlankItem(1L));
                     }
-                    blueData.remove(data);
-                    blueData.add((BaseSolariumData)BlankItem.generateBlankItem(1L));
                 }
             });
             addCosm.setVisible(false);
